Extract widgets in their own files
