+++ src/ui/widgets/__init__.py
@@ -1,4 +1,8 @@
 from .profilerow import ProfileRow
 from .multiline import MultilineEntry
+from .barchart import BarChartBox
+from .comborow import ComboRowHelper
+from .copybox import CopyBox
+from .file import File
 
-__all__ = ["ProfileRow", "MultilineEntry"]
+__all__ = ["ProfileRow", "MultilineEntry", "BarChartBox", "ComboRowHelper", "CopyBox", "File"]
