+++ src/window.py
@@ -9,7 +9,8 @@ import base64
 from .ui.profile import ProfileDialog
 from .llm import LLMHandler
 from .ui.presentation import PresentationWindow
-from .ui.gtkobj import File, CopyBox, BarChartBox, apply_css_to_widget
+from .ui.widgets import File, CopyBox, BarChartBox
+from .ui import apply_css_to_widget
 from .ui.widgets import MultilineEntry, ProfileRow
 from .constants import AVAILABLE_LLMS, AVAILABLE_PROMPTS, PROMPTS, AVAILABLE_TTS, AVAILABLE_STT
 from gi.repository import Gtk, Adw, Pango, Gio, Gdk, GObject, GLib, GdkPixbuf
